Refactor FCE Table to reduce duplication
